Archive all old bots
