+++ TeamCode/Archived/proto/debug/Noop.kt
