--- TeamCode/Joker/src/main/res/raw/readme.md
@@ -1,1 +1,0 @@
-Place your sound files in this folder to use them as project resources.
