--- TeamCode/Proto/build.gradle
@@ -1,13 +1,0 @@
-apply from: "../build.common.gradle"
-android {
-    namespace = 'org.firstinspires.ftc.teamcode'
-}
-
-buildscript {
-    repositories {
-        mavenCentral()
-        maven { url "https://repo.dairy.foundation/releases" }
-    }
-    dependencies { classpath "dev.frozenmilk:Load:0.2.1" }
-}
-apply plugin: "dev.frozenmilk.sinister.sloth.load"
