--- TeamCode/Proto/src/main/AndroidManifest.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!-- Note: the actual manifest file used in your APK merges this file with contributions
-     from the modules on which your app depends (such as FtcRobotController, etc).
-     So it won't ultimately be as empty as it might here appear to be :-) -->
-<manifest package="org.firstinspires.ftc.teamcode">
-    <application />
-</manifest>
