+++ settings.gradle
@@ -5,19 +5,11 @@ include ':Common'
 project(":Common").projectDir = new File("TeamCode/Common")
 include ':Cellphone'
 project(":Cellphone").projectDir = new File("TeamCode/Cellphone")
-include ':Proto'
-project(":Proto").projectDir = new File("TeamCode/Proto")
-include ':Vance'
-project(":Vance").projectDir = new File("TeamCode/Vance")
-include ':Joker'
-project(":Joker").projectDir = new File("TeamCode/Joker")
-include ':Jonas'
-project(":Jonas").projectDir = new File("TeamCode/Jonas")
-include ':IDK'
-project(":IDK").projectDir = new File("TeamCode/IDK")
 include ':Scout'
 project(":Scout").projectDir = new File("TeamCode/Scout")
 include ':Damon'
 project(":Damon").projectDir = new File("TeamCode/Damon")
+include ':Jonas'
+project(":Jonas").projectDir = new File("TeamCode/Jonas")
 include ':Lilbro5000'
 project(":Lilbro5000").projectDir = new File ("TeamCode/Lilbro5000")
